# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Milicic FireControl 365 API
  description: API REST corporativa para la gestión de inspecciones periódicas de extintores contra incendio, trazabilidad bajo norma IRAM 3517-2 y exportación a Microsoft 365.
  version: 2.5.0
  contact:
    name: Departamento de Higiene y Seguridad Laboral - Milicic S.A.
servers:
  - url: http://localhost:3000
    description: Servidor Local de Desarrollo
  - url: https://matafuegos.milicic.com.ar
    description: Servidor de Producción (Dokploy / Docker)
paths:
...
  /api/inspections:
    get:
      summary: Consultar historial de inspecciones
      parameters:
        - name: round_id
          in: query
          schema:
            type: integer
        - name: extinguisher_code
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Historial de inspecciones auditables

    post:
      summary: Registrar control mensual de extintor (IRAM 3517-2)
      description: Registra una inspección inmutable. En caso de fallas, genera automáticamente el caso de anomalía.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionInput'
      responses:
        '201':
          description: Inspección guardada con éxito

    put:
      summary: Intento de edición de inspección (Inmutable)
      responses:
        '405':
          description: Las inspecciones normativas son inmutables y no admiten modificación

    delete:
      summary: Intento de eliminación de inspección (Inmutable)
      responses:
        '405':
          description: Las inspecciones normativas son inmutables y no admiten eliminación
...
components:
  schemas:
...
    InspectionInput:
      type: object
      required:
        - extinguisher_code
        - inspector_name
        - passed
      properties:
        extinguisher_code:
          type: string
          example: MF-001
        inspector_name:
          type: string
          example: Santiago Amaya
        passed:
          type: integer
          enum: [0, 1]
          example: 1
        check_location:
          type: integer
          enum: [0, 1]
        check_pressure:
          type: integer
          enum: [0, 1]
        check_seal:
          type: integer
          enum: [0, 1]
        check_physical:
          type: integer
          enum: [0, 1]
        check_signage:
          type: integer
          enum: [0, 1]
        check_card:
          type: integer
          enum: [0, 1]
        observations:
          type: string
        photo_url:
          type: string